+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Version 1.4 (not yet released)
 -----------
 
 * Changed the implementation of the civicrm_container hook.
-* Added action to upload files
+* Added action to upload files for a contact
+* Added action to upload files for an event
+* Added action to create an organization
 * Added action to retrieve the labels from option values
 
 Version 1.3